Sketch out tesselation of SimpleShapes into Geometries using lyon
