+++ engine/monet/Cargo.toml
@@ -8,7 +8,8 @@ build = "src/build.rs"
 glium = "0.17.0"
 clippy = { version = "0.0.175", optional = true }
 fnv = "1.0.3"
-itertools = "0.7.4"
+itertools = "0.7.6"
+lyon_tessellation = "0.9.1"
 
 [dependencies.kay]
 path = "../kay"
