+++ engine/monet/src/lib.rs
@@ -11,16 +11,18 @@ extern crate compact;
 extern crate compact_macros;
 extern crate fnv;
 extern crate itertools;
+extern crate lyon_tessellation;
 
 mod geometry;
+mod geometry_actors;
 mod renderer;
 mod render_context;
 mod scene;
 
 pub use glium::backend::glutin::Display;
 
-pub use geometry::{Geometry, Batch, Vertex, Instance, Grouper, GrouperID, GrouperIndividual,
-                   GrouperIndividualID};
+pub use geometry::{Geometry, Batch, Vertex, Instance};
+pub use geometry_actors::{Grouper, GrouperID, GrouperIndividual, GrouperIndividualID};
 pub use renderer::{setup, Renderer, RendererID, Renderable, RenderableID, TargetProvider,
                    TargetProviderID, Movement, EyeListener, EyeListenerID, ProjectionRequester,
                    ProjectionRequesterID};
